refactor(app) : correct wrong path
