+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 // import { Navbar } from '@/components';
 import { ArtworkCarousel } from '../components/home/artworkCarousel';
 import { CuratedArtWorkAttribute } from '@/model/interface/curatedArtwork-types';
-import { getWeekArtWorkData } from '@/server-action/api.backend';
+import { getWeekArtWorkData } from '../server-action/backend/painting/api';
 import ScrollTriggerNavigator from '../components/quiz/ScrollTriggerNavigator';
 import { getQuizIDByContext } from './lib/api.quiz';
 import { Quiz } from '../model/interface/quiz';
