+++ src/components/search/PaintingCardGrid.tsx
@@ -5,10 +5,10 @@ import HoverCard from "../HoverCard";
 import { PreviewPainting } from '../PreviewPainting';
 import { Modal } from "../Modal";
 import { PaintingDetailView } from "../PaintingDetailView";
-import { findPainting, getPainting } from "../../server-action/api.backend";
-import { FindPaintingResult } from '@/server-action/dto';
 import { useSearchParams } from "next/navigation";
 import { throttle } from "../../util/optimization";
+import { FindPaintingResult } from "../../server-action/backend/painting/dto";
+import { findPainting, getPainting } from "../../server-action/backend/painting/api";
 
 interface PaintingCardGridProps  {
     findResult : FindPaintingResult;
